Crate runwell_wasm

Structs

Const
A Wasm translated Runwell constant value.

Error
An error that occurred while parsing a Wasm input and building up the module for it.

Export
An export item from a Wasm export section.

FunctionType
A function type.

GlobalVariable
A global variable declaration.

ImportName
The module and field name of an imported Wasm entity.

InitExpr
A parsed and validated Wasm constant initializer expression.

LinearMemoryDecl
A Wasm linear memory declaration.

MemoryDataInit
A data section item coming from the Wasm parser.

TableDecl
A Wasm table declaration.

Type
A Wasm translated Runwell type.

UnexpectedWasmPayload
An unexpected [wasmparser::Payload] has been encountered.

Enums

ErrorKind
Any kind of error that might occur while parsing a Wasm input binary.

ExportError
An error upon parsing, validating or operating on Wasm exports.

ExportItem
An export kind and a reference to the entity it exports.

ExportKind
Supported kinds of exports.

ImportError
An error upon parsing, validating or operating on Wasm imports.

InitExprError
An error that can occur upon parsing a global initializer expression.

MemoryError
An error that can occur upon parsing and validating linear memory.

PrimitiveError
An error that occurred while operating on Wasm primitives.

ReadError
Errors returned by Read::read.

SectionError
A general error that might occur while parsing Wasm sections.

TableError
An error that might occur while parsing or validating tables or table elements.

TranslateError
An error that occurred while translating from Wasm to Runwell IR.

UnsupportedTypeDef
Encountered an unsupported Wasm type definition.

UnsupportedWasmSection
Encountered an unsupported binary Wasm section.

Traits

Read
The Read trait allows for reading bytes from a source.

Functions

parse
Parses the binary WebAssembly (Wasm) bytes given through reader.